Relocate Logger class unto utils
